--- desktop/src-tauri/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "openjarvis-desktop"
-version = "0.1.0"
-description = "OpenJarvis Desktop — Native AI assistant with energy monitoring, trace debugging, and learning visualization"
-edition = "2021"
-license = "MIT"
-
-[build-dependencies]
-tauri-build = { version = "2", features = [] }
-
-[dependencies]
-tauri = { version = "2", features = ["tray-icon"] }
-tauri-plugin-notification = "2"
-tauri-plugin-shell = "2"
-tauri-plugin-global-shortcut = "2"
-tauri-plugin-autostart = "2"
-tauri-plugin-updater = "2"
-tauri-plugin-single-instance = "2"
-tauri-plugin-process = "2"
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-reqwest = { version = "0.12", features = ["json", "multipart"] }
-tokio = { version = "1", features = ["full"] }
-
-[target.'cfg(target_os = "macos")'.dependencies]
-objc = "0.2"
-dispatch = "0.2"
-
-[features]
-default = ["custom-protocol"]
-custom-protocol = ["tauri/custom-protocol"]
