--- desktop/src-tauri/src/main.rs
@@ -1,6 +1,0 @@
-// Prevents additional console window on Windows in release
-#![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
-
-fn main() {
-    openjarvis_desktop::run();
-}
