--- desktop/vite.config.ts
@@ -1,20 +1,0 @@
-import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
-
-export default defineConfig({
-  plugins: [react()],
-  build: {
-    outDir: 'dist',
-    emptyOutDir: true,
-  },
-  server: {
-    port: 5173,
-    strictPort: true,
-    proxy: {
-      '/v1': 'http://localhost:8000',
-      '/health': 'http://localhost:8000',
-    },
-  },
-  clearScreen: false,
-  envPrefix: ['VITE_', 'TAURI_'],
-});
