+++ frontend/src-tauri/binaries/.gitignore
