+++ frontend/src-tauri/capabilities/default.json
