+++ frontend/src/components/Desktop/EnergyDashboard.tsx
