+++ frontend/src/components/Desktop/LearningCurve.tsx
