+++ frontend/src/components/Desktop/MemoryBrowser.tsx
