+++ frontend/src/components/Desktop/TraceDebugger.tsx
