+++ frontend/tsconfig.json
@@ -21,5 +21,6 @@
       "@/*": ["./src/*"]
     }
   },
-  "include": ["src"]
+  "include": ["src"],
+  "exclude": ["src/components/Desktop"]
 }
